input test
